libwasi: move a few things to wasi_poll.c
