+++ libwasi/CMakeLists.txt
@@ -4,6 +4,7 @@ set(lib_wasi_sources
 	"wasi.c"
 	"wasi_fdtable.c"
 	"wasi_path.c"
+	"wasi_poll.c"
 	"wasi_subr.c"
 	"wasi_host_pathop.c"
 	"wasi_host_fdop.c"
